Remove macOS CLT from toolchain setup
Theoretically these should be picked up by the calls above that fetch
the actual include paths given the current compiler. At some point when
I added this it seemed like that wasn't the case if you had both Xcode
and the CLT installed, but I cannot reproduce that case to debug. This
otherwise massively increased the toolchain setup time when generating
the module map.